put error into own file
